backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
scipy==1.13.0
reportlab==4.2.0
matplotlib==3.8.4
Pillow==10.3.0
